Refactoring in Collaboration Service:
- Refactor comments in collab service
- Move interface from controller folder to types folder
- Refactor README.md
- Remove unused variable in queue: MATCH_FOUND
