+++ services/collaboration/src/events/queues.ts
@@ -2,7 +2,6 @@
  * Enum for queues
  */
 export enum Queues {
-    MATCH_FOUND = 'MATCH_FOUND',
     QUESTION_FOUND = 'QUESTION_FOUND',
     COLLAB_CREATED = 'COLLAB_CREATED',
     MATCH_FAILED = 'MATCH_FAILED',
