+++ services/collaboration/src/routes/index.ts
@@ -1,5 +1,6 @@
 import express from 'express';
 import { getHealth } from '../controllers';
+
 const router = express.Router();
 
 router.get('/ht', getHealth);
